add: avatar for userprofile
