--- src/main/java/by/itstep/controller/RegistrationController.java
@@ -1,79 +1,0 @@
-package by.itstep.controller;
-
-import by.itstep.model.jpa.User;
-import by.itstep.service.UserService;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.util.StringUtils;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import javax.validation.Valid;
-import java.util.Map;
-
-@Controller
-public class RegistrationController {
-
-    private final UserService userService;
-
-    public RegistrationController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @GetMapping("/registration")
-    public String registration(Model model) {
-        model.addAttribute("message", "");
-        return ("/registration");
-    }
-
-    @PostMapping("/registration")
-    public String addUser(
-            @RequestParam("password2") String passwordConfirm,
-            @Valid User user,
-            BindingResult bindingResult,
-            Model model
-    ) {
-        if (userService.isConfirmEmpty(passwordConfirm)){
-            model.addAttribute("password2Error", "Password confirmation can't be empty");
-        }
-
-        if (userService.isPasswordDifferent(passwordConfirm, user)){
-            bindingResult.addError(new FieldError("user", "password", "Password are different"));
-            model.addAttribute("passwordError", "Password are different");
-            return "registration";
-        }
-
-        if (userService.isConfirmEmpty(passwordConfirm) || bindingResult.hasErrors()) {
-            Map<String, String> errors = ControllerUtils.getErrors(bindingResult);
-
-            model.mergeAttributes(errors);
-
-            return "registration";
-        }
-
-        if (!userService.addUser(user)) {
-            model.addAttribute("usernameError", "User exists!");
-            return "registration";
-        }
-        return "redirect:/login";
-    }
-
-    @GetMapping("/activate/{code}")
-    public String activate(Model model, @PathVariable String code){
-        boolean isActivated = userService.activateUser(code);
-
-        if (isActivated) {
-            model.addAttribute("messageType", "success");
-            model.addAttribute("message", "User successfully activated");
-        } else {
-            model.addAttribute("messageType", "danger");
-            model.addAttribute("message", "Activation code is not found");
-        }
-
-        return "login";
-    }
-}
