+++ src/main/java/by/itstep/lms/common/Messages.java
@@ -1,0 +1,14 @@
+package by.itstep.lms.common;
+
+public class Messages {
+
+    public static final String ACTIVATE_ACCOUNT_MESSAGE = "Hello, Admin! \n" +
+            "Someone with username %s and email %s want to registration in Learn management system." +
+            "Please visit next link to activate this account: http://localhost:8080/activate/%s";
+
+    //subject messages
+    public static final String ACTIVATION_MASSAGE = "Activation account";
+
+    //front messages
+    public static final String NONE = "<none>";
+}
