+++ src/main/java/by/itstep/lms/config/DownloadInitializer.java
@@ -1,0 +1,24 @@
+package by.itstep.lms.config;
+
+import org.springframework.web.WebApplicationInitializer;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+
+import javax.servlet.ServletContext;
+import javax.servlet.ServletRegistration;
+
+public class DownloadInitializer implements WebApplicationInitializer {
+
+    public void onStartup(ServletContext container) {
+
+        AnnotationConfigWebApplicationContext ctx = new AnnotationConfigWebApplicationContext();
+        ctx.register(DownloadConfig.class);
+        ctx.setServletContext(container);
+
+        ServletRegistration.Dynamic servlet = container.addServlet("dispatcher", new DispatcherServlet(ctx));
+
+        servlet.setLoadOnStartup(1);
+        servlet.addMapping("/");
+    }
+
+}
