+++ src/main/java/by/itstep/lms/entity/Image.java
@@ -1,0 +1,28 @@
+package by.itstep.lms.entity;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+
+import javax.persistence.Id;
+import javax.persistence.Lob;
+
+import java.util.Base64;
+
+import static javax.persistence.GenerationType.IDENTITY;
+
+@Data
+@Entity
+public class Image {
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Integer id;
+    private String name;
+    @Lob
+    private byte[] data;
+
+    public String getPicture(){
+        return Base64.getEncoder().encodeToString(data);
+    }
+}
