+++ src/main/java/by/itstep/lms/entity/User.java
@@ -1,0 +1,108 @@
+package by.itstep.lms.entity;
+
+import by.itstep.lms.model.Role;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import org.hibernate.annotations.Fetch;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import javax.persistence.ElementCollection;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.CollectionTable;
+import javax.persistence.Enumerated;
+import javax.persistence.EnumType;
+import javax.persistence.ManyToMany;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
+
+import static javax.persistence.CascadeType.REMOVE;
+import static javax.persistence.FetchType.EAGER;
+import static javax.persistence.GenerationType.IDENTITY;
+import static org.hibernate.annotations.FetchMode.SELECT;
+
+@Data
+@Entity
+@Table(name = "usr")
+@EqualsAndHashCode(exclude = "groups")
+public class User implements UserDetails {
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long userId;
+
+    @NotBlank(message = "Username can't be empty")
+    private String username;
+    @NotBlank(message = "Password can't be empty")
+    private String password;
+    private boolean active;
+
+    @Email(message = "Email isn't correct")
+    @NotBlank(message = "Email can't be empty")
+    private String email;
+    private String activationCode;
+
+    @OneToOne
+    @JoinColumn(name = "image_id")
+    private Image avatar;
+
+    @ElementCollection(targetClass = Role.class, fetch = EAGER)
+    @CollectionTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"))
+    @Enumerated(EnumType.STRING)
+    private Set<Role> roles;
+
+    @Fetch(value = SELECT)
+    @ManyToMany(fetch = EAGER, cascade = REMOVE)
+    @JoinTable(
+            name = "user_group",
+            joinColumns = {@JoinColumn(name = "user_id")},
+            inverseJoinColumns = {@JoinColumn (name = "group_id")}
+    )
+    private Set<Group> groups = new HashSet<>();
+
+    public boolean isAdmin(){
+        return roles.contains(Role.ADMIN);
+    }
+
+    public boolean isTeacher(){
+        return roles.contains(Role.TEACHER);
+    }
+
+    public boolean isUser(){
+        return roles.contains(Role.USER);
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return getRoles();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return isActive();
+    }
+}
