+++ src/main/java/by/itstep/lms/model/Role.java
@@ -1,0 +1,12 @@
+package by.itstep.lms.model;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum Role implements GrantedAuthority {
+    USER, ADMIN, TEACHER;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+}
