+++ src/main/java/by/itstep/lms/model/dto/IdDto.java
@@ -1,0 +1,9 @@
+package by.itstep.lms.model.dto;
+
+import lombok.Data;
+
+@Data
+public class IdDto {
+    private Long userId;
+    private Long groupId;
+}
