+++ src/main/java/by/itstep/lms/repository/GroupRepository.java
@@ -1,0 +1,10 @@
+package by.itstep.lms.repository;
+
+import by.itstep.lms.entity.Group;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface GroupRepository extends CrudRepository<Group, Long> {
+    Optional<Group> findByGroupId(Long id);
+}
