+++ src/main/java/by/itstep/lms/repository/HomeworkRepository.java
@@ -1,0 +1,7 @@
+package by.itstep.lms.repository;
+
+import by.itstep.lms.entity.Homework;
+import org.springframework.data.repository.CrudRepository;
+
+public interface HomeworkRepository extends CrudRepository<Homework, Long> {
+}
