+++ src/main/java/by/itstep/lms/repository/ImageRepository.java
@@ -1,0 +1,9 @@
+package by.itstep.lms.repository;
+
+import by.itstep.lms.entity.Image;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ImageRepository extends JpaRepository<Image, String> {
+    Image findFirstByName(String name);
+    Image findFirstById(int id);
+}
