+++ src/main/java/by/itstep/lms/repository/SolutionRepository.java
@@ -1,0 +1,7 @@
+package by.itstep.lms.repository;
+
+import by.itstep.lms.entity.Solution;
+import org.springframework.data.repository.CrudRepository;
+
+public interface SolutionRepository extends CrudRepository<Solution, Long> {
+}
