+++ src/main/java/by/itstep/lms/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package by.itstep.lms.repository;
+
+import by.itstep.lms.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+
+public interface UserRepository extends JpaRepository<User, Long>{
+    User findByUsername(String login);
+
+    User findByActivationCode(String code);
+
+    Optional<User> findByUserId(Long userId);
+}
