--- src/main/java/by/itstep/model/Role.java
@@ -1,12 +1,0 @@
-package by.itstep.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role implements GrantedAuthority {
-    USER, ADMIN, TEACHER;
-
-    @Override
-    public String getAuthority() {
-        return name();
-    }
-}
