--- src/main/java/by/itstep/model/dto/IdDto.java
@@ -1,9 +1,0 @@
-package by.itstep.model.dto;
-
-import lombok.Data;
-
-@Data
-public class IdDto {
-    private Long userId;
-    private Long groupId;
-}
