--- src/main/java/by/itstep/repository/GroupRepository.java
@@ -1,10 +1,0 @@
-package by.itstep.repository;
-
-import by.itstep.model.jpa.Group;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface GroupRepository extends CrudRepository<Group, Long> {
-    Optional<Group> findByGroupId(Long id);
-}
