--- src/main/java/by/itstep/repository/HomeworkRepository.java
@@ -1,7 +1,0 @@
-package by.itstep.repository;
-
-import by.itstep.model.jpa.Homework;
-import org.springframework.data.repository.CrudRepository;
-
-public interface HomeworkRepository extends CrudRepository<Homework, Long> {
-}
