--- src/main/java/by/itstep/repository/SolutionRepository.java
@@ -1,7 +1,0 @@
-package by.itstep.repository;
-
-import by.itstep.model.jpa.Solution;
-import org.springframework.data.repository.CrudRepository;
-
-public interface SolutionRepository extends CrudRepository<Solution, Long> {
-}
