+++ src/main/resources/config/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mariadb://localhost:3306/lms
 spring.datasource.username=root
-spring.datasource.password=use database passsword here
+spring.datasource.password=use database password here
 spring.jpa.generate-ddl=false
 spring.jpa.show-sql=false
 spring.jpa.hibernate.ddl-auto=validate
