+++ src/main/resources/templates/parts/common.ftlh
@@ -14,7 +14,6 @@
     </head>
     <body>
     <#include "navbar.ftlh">
-    <#include "footer.ftlh">
     <div class="container mt-5">
     <#nested>
     </div>
@@ -25,5 +24,6 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
     </body>
+    <#include "footer.ftlh">
     </html>
 </#macro>
